+++ src/Calendar.js
@@ -1,5 +1,7 @@
 import React, { Component } from 'react';
+import moment from 'moment';
 import PropTypes from 'prop-types';
+import cx from 'classnames';
 import createDateObjects from './createDateObjects';
 
 export default class Calendar extends Component {
@@ -19,12 +21,28 @@ export default class Calendar extends Component {
     /** Called when some of the navigation controls are clicked */
     onChangeMonth: PropTypes.func,
     /** Called when a date is clicked */
-    onPickDate: PropTypes.func
+    onPickDate: PropTypes.func,
+    /** classname for div wrapping the whole calendar */
+    containerClassName: PropTypes.string,
+    /** classname for the div wrapping the grid */
+    contentClassName: PropTypes.string
   };
 
   static defaultProps = {
     weekOffset: 0,
-    renderDay: day => day.format('YYYY-MM-D'),
+    renderDay: ({ day, classNames, onPickDate }) => (
+      <div
+        key={day.format()}
+        className={cx(
+          'Calendar-grid-item',
+          day.isSame(moment(), 'day') && 'Calendar-grid-item--current',
+          classNames
+        )}
+        onClick={e => onPickDate(day)}
+      >
+        {day.format('D')}
+      </div>
+    ),
     renderHeader: ({ date, onPrevMonth, onNextMonth }) => (
       <div className="Calendar-header">
         <button onClick={onPrevMonth}>«</button>
@@ -58,28 +76,22 @@ export default class Calendar extends Component {
       weekOffset,
       renderDay,
       renderHeader,
-      onPickDate
+      onPickDate,
+      contentClassName,
+      containerClassName
     } = this.props;
 
-    const { handlePrevMonth, handleNextMonth } = this;
-
     return (
-      <div className="Calendar">
+      <div className={cx('Calendar', containerClassName)}>
         {renderHeader({
           date,
-          onPrevMonth: handlePrevMonth,
-          onNextMonth: handleNextMonth
+          onPrevMonth: this.handlePrevMonth,
+          onNextMonth: this.handleNextMonth
         })}
-        <div className="Calendar-grid">
-          {createDateObjects(date, weekOffset).map((day, i) => (
-            <div
-              key={`day-${i}`}
-              className={`Calendar-grid-item ${day.classNames || ''}`}
-              onClick={e => onPickDate(day.day)}
-            >
-              {renderDay(day.day)}
-            </div>
-          ))}
+        <div className={cx('Calendar-grid', contentClassName)}>
+          {createDateObjects(date, weekOffset).map((day, i) =>
+            renderDay({ ...day, onPickDate })
+          )}
         </div>
       </div>
     );